style: 폴더 링크 카드 컴포넌트 반응형 스타일
