+++ src/components/FolderLinks/style.css
@@ -1,0 +1,56 @@
+.folder-link-box {
+  width: 34rem;
+  height: 33.4rem;
+  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
+  border-radius: 2rem;
+  cursor: pointer;
+  overflow: hidden;
+}
+
+.link-image-box {
+  overflow: hidden;
+  width: 34rem;
+  height: 22rem;
+}
+
+.link-image {
+  object-fit: cover;
+  width: 100%;
+  height: 100%;
+  transition: all 0.4s;
+}
+
+.link-image:hover {
+  transform: scale(1.3);
+}
+
+.link-description-box {
+  margin: 0 2rem;
+}
+
+.written-date {
+  color: var(--gray60);
+  font-size: 1.3rem;
+  margin: 1.1rem 0;
+}
+
+.description {
+  text-overflow: ellipsis;
+  white-space: nowrap;
+  overflow: hidden;
+  font-size: 1.6rem;
+  line-height: 2.4rem;
+  margin: 1rem 0;
+}
+
+.format-date {
+  font-size: 1.4rem;
+}
+
+/* 모바일 */
+@media (max-width: 767px) {
+  .folder-link-box {
+    width: 32.5rem;
+    height: 32.7rem;
+  }
+}
