+++ src/pages/SharedPage/index.js
@@ -2,20 +2,31 @@ import Nav from "../../components/Nav";
 import FolderLinks from "../../components/FolderLinks";
 import Footer from "../../components/Footer";
 import "./style.css";
+import search from "../../assets/search.svg";
 
 const SharedPage = ({ user, folderObj }) => {
   const { folder = {} } = folderObj;
   const { owner = {}, links = [] } = folder;
 
   return (
-    <main>
+    <>
       <Nav user={user} />
-      <img src={owner.profileImageSource} alt={owner.name} />
-      <div>@{owner.name}</div>
-      <div>{folder.name}</div>
-      <FolderLinks links={links} />
+      <main>
+        <section className="folder-info">
+          <img className="folder-profile-image" src={owner.profileImageSource} alt={owner.name} />
+          <div className="folder-profile-name">@{owner.name}</div>
+          <div className="folder-name">{folder.name}</div>
+        </section>
+        <section className="folder-links">
+          <div className="link-search-bar">
+            <img className="link-search-icon" src={search} alt="검색 아이콘" />
+            <input className="link-search" placeholder="링크를 검색해 보세요." />
+          </div>
+          <FolderLinks className="SharedPage-folder-links" links={links} />
+        </section>
+      </main>
       <Footer />
-    </main>
+    </>
   );
 };
