+++ src/pages/SharedPage/style.css
@@ -1,0 +1,86 @@
+.folder-info {
+  background: var(--background);
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+  padding: 2rem 0 6rem;
+}
+
+.folder-profile-image {
+  width: 6rem;
+  height: 6rem;
+}
+
+.folder-profile-name {
+  font-size: 1.6rem;
+  margin-top: 1.2rem;
+  margin-bottom: 2rem;
+}
+
+.folder-name {
+  font-size: 4rem;
+  font-weight: 600;
+}
+
+.folder-links {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-direction: column;
+}
+
+.link-search-bar {
+  position: relative;
+  padding: 4rem 0;
+}
+
+.link-search-icon {
+  position: absolute;
+  top: 5.8rem;
+  left: 1.6rem;
+  width: 1.6rem;
+  height: 1.6rem;
+}
+
+.link-search {
+  border: none;
+  background: #f5f5f5;
+  padding: 1.5rem 1.6rem 1.5rem 4rem;
+  width: 106rem;
+  height: 5.4rem;
+  border-radius: 1rem;
+}
+
+.SharedPage-folder-links {
+  padding: 0;
+  display: grid;
+  grid-template: repeat(3, 1fr) / repeat(3, 1fr);
+  gap: 2rem;
+}
+
+/* 태블릿 */
+@media (max-width: 1124px) {
+  .link-search {
+    width: 70rem;
+  }
+
+  .SharedPage-folder-links {
+    grid-template: repeat(2, 1fr) / repeat(2, 1fr);
+  }
+}
+
+/* 모바일 */
+@media (max-width: 767px) {
+  .folder-links {
+    margin: 0 3.2rem;
+  }
+
+  .link-search {
+    width: 32.5rem;
+  }
+
+  .SharedPage-folder-links {
+    grid-template: repeat(1, 1fr) / repeat(1, 1fr);
+  }
+}
