+++ src/styles/reset.css
@@ -18,6 +18,10 @@ a {
   cursor: pointer;
 }
 
+li {
+  list-style: none;
+}
+
 button {
   border: none;
   cursor: pointer;
